requirements.txt
Flask==3.0.0
python-binance==1.0.19
requests==2.31.0

# Optional: for testing
# pytest==7.4.3

